Finish api controller refactoring;
