+++ app/Http/Controllers/Api/PrivateApi/User/UserProfileApiController.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace app\Http\Controllers\Api\PrivateApi\User;
+namespace App\Http\Controllers\Api\PrivateApi\User;
 
 
 class UserProfileApiController
