+++ app/Http/Controllers/Api/PrivateApi/User/UserSettingsApiController.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace app\Http\Controllers\Api\PrivateApi\User;
+namespace App\Http\Controllers\Api\PrivateApi\User;
 
 
 class UserSettingsApiController
