+++ routes/papi.php
@@ -11,29 +11,29 @@
 |
 */
 
-Route::group(['namespace' => 'Api', 'middleware' => 'api-auth'], function () {
+Route::group(['namespace' => 'Api\PrivateApi', 'middleware' => 'api-auth'], function () {
 	$r = config('papi.resources');
 
 	// Courses
-	Route::get("{$r['courses']}/{id}", 'CoursesApiController@get');
+	Route::get("{$r['courses']}/{id}", 'Course\CoursesApiController@get');
 
 	// Lessons
-	Route::get("{$r['lessons']}/{id}", 'LessonsApiController@get');
+	Route::get("{$r['lessons']}/{id}", 'Course\LessonsApiController@get');
 
 	// Screens
-	Route::get("{$r['screens']}/{id}", 'ScreensApiController@get');
-
-	// Users
-	Route::get("{$r['users']}/{id}", 'UsersApiController@get');
-	Route::put("{$r['users']}/{id}", 'UsersApiController@put');
-
-	Route::get("{$r['users']}/{id}/{$r['user-profile']}", 'UsersApiController@getUserProfile');
-	Route::put("{$r['users']}/{id}/{$r['user-profile']}", 'UsersApiController@putUserProfile');
-
-	Route::post("{$r['users']}/{id}/{$r['user-avatar']}", 'UsersApiController@putUserAvatar');
+	Route::get("{$r['screens']}/{id}", 'Course\ScreensApiController@get');
 
 	// Editions
-	Route::get("{$r['editions']}/{id}", 'EditionsApiController@get');
+	Route::get("{$r['editions']}/{id}", 'Course\EditionsApiController@get');
+
+	// Users
+	Route::get("{$r['users']}/{id}", 'User\UsersApiController@get');
+	Route::put("{$r['users']}/{id}", 'User\UsersApiController@put');
+
+	Route::get("{$r['users']}/{id}/{$r['user-profile']}", 'User\UsersApiController@getUserProfile');
+	Route::put("{$r['users']}/{id}/{$r['user-profile']}", 'User\UsersApiController@putUserProfile');
+
+	Route::post("{$r['users']}/{id}/{$r['user-avatar']}", 'User\UsersApiController@putUserAvatar');
 
 	// Orders
 	Route::get("{$r['orders']}/{id}", 'OrdersApiController@get');
@@ -42,12 +42,12 @@ Route::group(['namespace' => 'Api', 'middleware' => 'api-auth'], function () {
 	Route::get("{$r['tags']}/{id}", 'TagsApiController@get');
 
 	// Questions
-	Route::get("{$r['questions']}/{id}", 'QuestionsApiController@get');
-	Route::post($r['questions'], 'QuestionsApiController@post');
+	Route::get("{$r['questions']}/{id}", 'Qna\QuestionsApiController@get');
+	Route::post($r['questions'], 'Qna\QuestionsApiController@post');
 
 	// Answers
-	Route::get("{$r['answers']}/{id}", 'AnswersApiController@get');
-	Route::post($r['answers'], 'AnswersApiController@post');
+	Route::get("{$r['answers']}/{id}", 'Qna\AnswersApiController@get');
+	Route::post($r['answers'], 'Qna\AnswersApiController@post');
 
 	// User Progress
 //	Route::get("{$r['users']}/{id}", 'CoursesApiController@get');
